backend-dev/requirements.txt
python-dotenv
flask
requests
Flask-Cors
